// semant_demo_frontend/src/generated/api/models/SearchFilter.ts
export interface SearchFilter {
    /**
     * 
     * @type {string}
     * @memberof SearchFilter
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof SearchFilter
     */
    name: string;
    /**
     * 
     * @type {FilterType}
     * @memberof SearchFilter
     */
    type: FilterType;
    /**
     * 
     * @type {string}
     * @memberof SearchFilter
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof SearchFilter
     */
    targetProperty: string;
    /**
     * 
     * @type {Array<NominalFilterValue>}
     * @memberof SearchFilter
     */
    values?: Array<NominalFilterValue> | null;
    /**
     * 
     * @type {MinValue}
     * @memberof SearchFilter
     */
    minValue?: MinValue | null;
    /**
     * 
     * @type {MaxValue}
     * @memberof SearchFilter
     */
    maxValue?: MaxValue | null;
}